+++ projects/rocfft/library/src/device/generator/stockham_gen.cpp
@@ -49,6 +49,7 @@ struct GeneratedLauncher
     GeneratedLauncher(StockhamKernel&                  kernel,
                       const std::string&               scheme,
                       const std::string&               pp_child_scheme,
+                      const unsigned int&              pp_threads_per_transform,
                       const std::vector<unsigned int>& pp_factors_curr,
                       const std::vector<unsigned int>& pp_factors_other,
                       const unsigned int&              pp_current_dim,
@@ -59,6 +60,7 @@ struct GeneratedLauncher
                       const std::string&               sbrc_transpose_type)
         : scheme(scheme)
         , pp_child_scheme(pp_child_scheme)
+        , pp_threads_per_transform(pp_threads_per_transform)
         , pp_factors_curr(pp_factors_curr)
         , pp_factors_other(pp_factors_other)
         , pp_current_dim(pp_current_dim)
@@ -79,6 +81,7 @@ struct GeneratedLauncher
 
     std::string               scheme;
     std::string               pp_child_scheme;
+    unsigned int              pp_threads_per_transform;
     std::vector<unsigned int> pp_factors_curr;
     std::vector<unsigned int> pp_factors_other;
     unsigned int              pp_current_dim;
@@ -139,6 +142,7 @@ struct GeneratedLauncher
         add_member("precision_type", std::to_string(precision_type));
         add_member("gcn_arch_name", quote_str(gcn_arch_name));
         add_member("pp_child_scheme", quote_str(pp_child_scheme));
+        add_member("pp_threads_per_transform", std::to_string(pp_threads_per_transform));
         add_member("pp_factors_curr", vec_to_list(pp_factors_curr));
         add_member("pp_factors_other", vec_to_list(pp_factors_other));
         add_member("pp_current_dim", std::to_string(pp_current_dim));
@@ -162,6 +166,7 @@ void make_launcher(const unsigned int&              precision_type,
                    StockhamKernel&                  kernel,
                    const std::string&               gcn_arch_name,
                    const std::string&               pp_child_scheme,
+                   const unsigned&                  pp_threads_per_transform,
                    const std::vector<unsigned int>& pp_factors_curr,
                    const std::vector<unsigned int>& pp_factors_other,
                    const unsigned int&              pp_current_dim,
@@ -174,6 +179,7 @@ void make_launcher(const unsigned int&              precision_type,
         generated_launchers.emplace_back(kernel,
                                          launcher.scheme,
                                          pp_child_scheme,
+                                         pp_threads_per_transform,
                                          pp_factors_curr,
                                          pp_factors_other,
                                          pp_current_dim,
@@ -268,6 +274,7 @@ void stockham_partial_pass_variants(const std::string&               kernel_name
                           kernelRR,
                           specs1.gcn_arch_name,
                           "CS_KERNEL_STOCKHAM_PP",
+                          params_1.pp_threads_per_transform,
                           params_1.pp_factors_curr,
                           params_1.pp_factors_other,
                           params_1.current_dim,
@@ -280,6 +287,7 @@ void stockham_partial_pass_variants(const std::string&               kernel_name
                           kernelCC,
                           specs2.gcn_arch_name,
                           "CS_KERNEL_STOCKHAM_PP_BLOCK_CC",
+                          params_2.pp_threads_per_transform,
                           params_2.pp_factors_curr,
                           params_2.pp_factors_other,
                           params_2.current_dim,
@@ -294,6 +302,7 @@ void stockham_partial_pass_variants(const std::string&               kernel_name
                           kernelCC,
                           specs1.gcn_arch_name,
                           "CS_KERNEL_STOCKHAM_PP_BLOCK_CC",
+                          params_1.pp_threads_per_transform,
                           params_1.pp_factors_curr,
                           params_1.pp_factors_other,
                           params_1.current_dim,
@@ -306,6 +315,7 @@ void stockham_partial_pass_variants(const std::string&               kernel_name
                           kernelRR,
                           specs2.gcn_arch_name,
                           "CS_KERNEL_STOCKHAM_PP",
+                          params_2.pp_threads_per_transform,
                           params_2.pp_factors_curr,
                           params_2.pp_factors_other,
                           params_2.current_dim,
@@ -358,6 +368,7 @@ void stockham_variants(const std::string&            kernel_name,
                       kernel,
                       specs.gcn_arch_name,
                       "CS_NONE",
+                      0,
                       std::vector<unsigned int>(),
                       std::vector<unsigned int>(),
                       0,
@@ -372,6 +383,7 @@ void stockham_variants(const std::string&            kernel_name,
                       kernel,
                       specs.gcn_arch_name,
                       "CS_NONE",
+                      0,
                       std::vector<unsigned int>(),
                       std::vector<unsigned int>(),
                       0,
@@ -424,6 +436,7 @@ void stockham_variants(const std::string&            kernel_name,
                       kernel,
                       specs.gcn_arch_name,
                       "CS_NONE",
+                      0,
                       std::vector<unsigned int>(),
                       std::vector<unsigned int>(),
                       0,
@@ -439,6 +452,7 @@ void stockham_variants(const std::string&            kernel_name,
                       kernel,
                       specs.gcn_arch_name,
                       "CS_NONE",
+                      0,
                       std::vector<unsigned int>(),
                       std::vector<unsigned int>(),
                       0,
@@ -452,6 +466,7 @@ void stockham_variants(const std::string&            kernel_name,
         launchers.emplace_back(fused2d,
                                specs.scheme,
                                "CS_NONE",
+                               0,
                                std::vector<unsigned int>(),
                                std::vector<unsigned int>(),
                                0,
@@ -493,6 +508,7 @@ void pp_params_rm(std::vector<unsigned int>& parent_length,
                   std::vector<unsigned int>& pp_factors2,
                   std::vector<unsigned int>& workgroup_size,
                   std::vector<unsigned int>& threads_per_transform,
+                  std::vector<unsigned int>& threads_per_transform_pp,
                   std::vector<bool>&         direct_to_from_reg)
 {
     std::reverse(parent_length.begin(), parent_length.end());
@@ -507,6 +523,7 @@ void pp_params_rm(std::vector<unsigned int>& parent_length,
         pp_factors1.swap(pp_factors2);
         std::reverse(workgroup_size.begin(), workgroup_size.end());
         std::reverse(threads_per_transform.begin(), threads_per_transform.end());
+        std::reverse(threads_per_transform_pp.begin(), threads_per_transform_pp.end());
         std::reverse(direct_to_from_reg.begin(), direct_to_from_reg.end());
     }
 }
@@ -625,6 +642,24 @@ void validate_pp_grid_params(const StockhamPartialPassParams& params_1,
         {
             throw std::runtime_error("invalid dimensions for CS_3D_PP");
         }
+
+        // Validate pp_threads_per_transform against threads_per_transform
+        if(params_1.pp_threads_per_transform > specs_1.threads_per_transform
+           || params_2.pp_threads_per_transform > specs_2.threads_per_transform)
+        {
+            throw std::runtime_error(
+                "CS_KERNEL_STOCKHAM_PP requires threads_per_transform_pp to be "
+                "equal or less than threads_per_transform");
+        }
+
+        // Validate pp_threads_per_transform against pp_factors_curr
+        if((params_1.pp_factors_curr.size() == 1 && params_1.pp_threads_per_transform > 1)
+           || (params_2.pp_factors_curr.size() == 1 && params_2.pp_threads_per_transform > 1))
+        {
+            throw std::runtime_error("CS_KERNEL_STOCKHAM_PP and CS_KERNEL_STOCKHAM_PP_BLOCK_CC "
+                                     "require threads_per_transform_pp to be 1 when "
+                                     "pp_factors has only one factor");
+        }
     }
     else
     {
@@ -672,7 +707,8 @@ int main()
         ++arg;
         scheme = *arg;
 
-        std::vector<unsigned int> parent_length, dims, pp_factors1, pp_factors2;
+        std::vector<unsigned int> parent_length, dims, threads_per_transform_pp, pp_factors1,
+            pp_factors2;
         if(scheme == "CS_3D_PP")
         {
             ++arg;
@@ -683,6 +719,9 @@ int main()
 
             ++arg;
             pp_factors1 = parse_uints_csv(*arg);
+
+            ++arg;
+            threads_per_transform_pp = parse_uints_csv(*arg);
 
             ++arg;
             dims = parse_uints_csv(*arg);
@@ -733,11 +772,16 @@ int main()
                          pp_factors2,
                          workgroup_size,
                          threads_per_transform,
+                         threads_per_transform_pp,
                          direct_to_from_reg);
 
             if(threads_per_transform.size() != 2)
                 throw std::runtime_error(
                     "CS_3D_PP requires two threads_per_transform configuration");
+
+            if(threads_per_transform_pp.size() != 2)
+                throw std::runtime_error(
+                    "CS_3D_PP requires two threads_per_transform_pp configuration");
 
             if(direct_to_from_reg.size() != 2)
                 throw std::runtime_error("CS_3D_PP requires two direct_to_from_reg configuration");
@@ -754,10 +798,18 @@ int main()
             specs2.threads_per_transform = threads_per_transform[1];
             specs2.wgs_is_derived        = true;
 
-            StockhamPartialPassParams pp_params_1(
-                parent_length, dims[0], off_dim, pp_factors1, pp_factors2);
-            StockhamPartialPassParams pp_params_2(
-                parent_length, dims[1], off_dim, pp_factors2, pp_factors1);
+            StockhamPartialPassParams pp_params_1(parent_length,
+                                                  threads_per_transform_pp[0],
+                                                  dims[0],
+                                                  off_dim,
+                                                  pp_factors1,
+                                                  pp_factors2);
+            StockhamPartialPassParams pp_params_2(parent_length,
+                                                  threads_per_transform_pp[1],
+                                                  dims[1],
+                                                  off_dim,
+                                                  pp_factors2,
+                                                  pp_factors1);
 
             validate_pp_length(pp_params_1, factors1);
             validate_pp_length(pp_params_2, factors2);