+++ projects/rocfft/library/src/device/generator/stockham_gen.h
@@ -49,15 +49,18 @@ struct StockhamGeneratorSpecs
 
     std::vector<unsigned int> factors;
     std::vector<unsigned int> factors2d;
+    std::vector<unsigned int> factors_pp;
     unsigned int              precision; // mapped from rocfft_precision
     std::string               gcn_arch_name;
     unsigned int              length;
+    unsigned int              length_pp;
     unsigned int              length2d = 0;
 
     unsigned int workgroup_size;
-    unsigned int threads_per_transform = 0;
-    bool         half_lds              = false;
-    bool         direct_to_from_reg    = false;
+    unsigned int threads_per_transform    = 0;
+    unsigned int threads_per_transform_pp = 0;
+    bool         half_lds                 = false;
+    bool         direct_to_from_reg       = false;
     // dimension of the kernel - 0 if the generated kernel accepts a
     // 'dim' argument at runtime; otherwise the dimension is
     // statically defined for the kernel
@@ -91,11 +94,13 @@ struct StockhamPartialPassParams
     StockhamPartialPassParams() = default;
 
     StockhamPartialPassParams(const std::vector<unsigned int>& parent_length,
+                              const unsigned int               pp_threads_per_transform,
                               const unsigned int               current_dim,
                               const unsigned int               off_dim,
                               const std::vector<unsigned int>& pp_factors_curr,
                               const std::vector<unsigned int>& pp_factors_other)
         : parent_length(parent_length)
+        , pp_threads_per_transform(pp_threads_per_transform)
         , current_dim(current_dim)
         , off_dim(off_dim)
         , pp_factors_curr(pp_factors_curr)
@@ -104,6 +109,7 @@ struct StockhamPartialPassParams
     }
 
     std::vector<unsigned int> parent_length;
+    unsigned int              pp_threads_per_transform;
     unsigned int              current_dim = 0;
     unsigned int              off_dim     = 0;
     std::vector<unsigned int> pp_factors_curr;