+++ projects/rocfft/library/src/device/kernel-generator.py
@@ -258,6 +258,7 @@ class FFTKernel(BaseNode):
         f += ', ' + str(self.function.meta.pp_child_scheme)
         f += ', ' + str(self.function.meta.pp_current_dim)
         f += ', ' + str(self.function.meta.pp_off_dim)
+        f += ', ' + str(self.function.meta.pp_threads_per_transform)
         pp_factors_curr = getattr(self.function.meta, 'pp_factors_curr', None)
         if pp_factors_curr is not None:
             f += ', {' + cjoin(pp_factors_curr) + '}'
@@ -518,13 +519,21 @@ def list_3d_partial_pass_kernels():
 
 
 def default_runtime_compile(kernels, default_val):
-    '''Returns a copy of input kernel list with a default value for runtime_compile.'''
+    """Returns a copy of input kernel list with a default value for runtime_compile.
+       Validates that AOT kernels only use gfx-generic as architecture."""
 
-    return [
-        k if hasattr(k, 'runtime_compile') else NS(**k.__dict__,
-                                                   runtime_compile=default_val)
-        for k in kernels
-    ]
+    for k in kernels:
+        if not hasattr(k, 'runtime_compile'):
+            k.runtime_compile = default_val
+
+        if k.runtime_compile is False:
+            # validate that gcn_arch_name is not gfx generic
+            if k.gcn_arch_name != config_arch.supported_arch.GFX_GENERIC.value:
+                err_msg = "Error: runtime_compile cannot be false for non gfx-generic architectures: \n"
+                print(err_msg + str(k))
+                sys.exit(1)
+
+    return kernels
 
 
 def generate_kernel_functions(precisions_type_dict, kernels, launchers_json):
@@ -559,6 +568,7 @@ def generate_kernel_functions(precisions_type_dict, kernels, launchers_json):
             direct_to_from_reg = launcher.direct_to_from_reg
             scheme = launcher.scheme
             pp_child_scheme = launcher.pp_child_scheme
+            pp_threads_per_transform = launcher.pp_threads_per_transform
             pp_factors_curr = launcher.pp_factors_curr
             pp_factors_other = launcher.pp_factors_other
             pp_current_dim = launcher.pp_current_dim
@@ -580,24 +590,26 @@ def generate_kernel_functions(precisions_type_dict, kernels, launchers_json):
             ]
 
             f = Function(arguments=ArgumentList(data, back),
-                         meta=NS(factors=factors,
-                                 length=length,
-                                 params=params,
-                                 precision=precision,
-                                 gcn_arch_name=gcn_arch_name,
-                                 runtime_compile=runtime_compile,
-                                 scheme=scheme,
-                                 workgroup_size=workgroup_size,
-                                 transforms_per_block=transforms_per_block,
-                                 threads_per_transform=tpt_list,
-                                 transpose=sbrc_transpose_type,
-                                 use_3steps_large_twd=use_3steps_large_twd,
-                                 lds_size_bytes=kernel.lds_size_bytes,
-                                 pp_child_scheme=pp_child_scheme,
-                                 pp_factors_curr=pp_factors_curr,
-                                 pp_factors_other=pp_factors_other,
-                                 pp_current_dim=pp_current_dim,
-                                 pp_off_dim=pp_off_dim))
+                         meta=NS(
+                             factors=factors,
+                             length=length,
+                             params=params,
+                             precision=precision,
+                             gcn_arch_name=gcn_arch_name,
+                             runtime_compile=runtime_compile,
+                             scheme=scheme,
+                             workgroup_size=workgroup_size,
+                             transforms_per_block=transforms_per_block,
+                             threads_per_transform=tpt_list,
+                             transpose=sbrc_transpose_type,
+                             use_3steps_large_twd=use_3steps_large_twd,
+                             lds_size_bytes=kernel.lds_size_bytes,
+                             pp_child_scheme=pp_child_scheme,
+                             pp_threads_per_transform=pp_threads_per_transform,
+                             pp_factors_curr=pp_factors_curr,
+                             pp_factors_other=pp_factors_other,
+                             pp_current_dim=pp_current_dim,
+                             pp_off_dim=pp_off_dim))
 
             if (scheme == 'CS_3D_PP'):
                 pp_kernel_functions.append(f)
@@ -704,6 +716,9 @@ def generate_kernels(precisions_dict, kernels, stockham_gen):
             # check for data specific to partial-pass 3D kernels
             if hasattr(k, 'dims'):
                 proc.stdin.write(" " + ','.join([str(f) for f in k.dims]))
+                proc.stdin.write(
+                    " " +
+                    ','.join([str(f) for f in k.threads_per_transform_pp]))
                 proc.stdin.write(" " +
                                  ','.join([str(f)
                                            for f in k.factors_pp[0]]) + " ")